refactor(convertOrganisations): Remove `moment` variable
This variable was added because I was going to add a feature where a
`.expiry` value was added to converted records that would have been
equal to 90 days after the `updated` value. I haven't added that feature
as of yet so I added a `TODO` where the logic should go and removed the
reference to `moment`.